refactor: split raw record iteration into microcrate
